+++ src/main/java/com/github/petha/correlationengine/model/IdfContainer.java
@@ -1,6 +1,6 @@
 package com.github.petha.correlationengine.model;
 
 public interface IdfContainer {
-    double getIdf(Integer index);
+    float getIdf(Integer index);
     void updateTermFrequency(SparseVector vector);
 }
